pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ballpushing_utils"
version = "0.1.0"
description = "Analysis toolkit and paper companion for 'Object manipulation and affordance learning in Drosophila' (Durrieu et al., 2026)."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Matthias Durrieu", email = "matthias.durrieu@epfl.ch" },
]
maintainers = [
    { name = "Matthias Durrieu", email = "matthias.durrieu@epfl.ch" },
]
keywords = [
    "drosophila",
    "behavior",
    "neuroscience",
    "object-manipulation",
    "affordance",
    "ball-pushing",
    "tracking",
    "sleap",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Visualization",
]

# Runtime dependencies. Version pins are intentionally loose so the package
# co-exists with recent SciPy/matplotlib/pandas stacks; strict pins used to
# reproduce the paper figures live in `requirements.txt`.
dependencies = [
    "numpy>=1.26",
    "pandas>=2.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "scikit-bio>=0.6",
    "statsmodels>=0.14",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "h5py>=3.10",
    "pyarrow>=14",
    "tqdm>=4.65",
    "PyYAML>=6.0",
    "numba>=0.59",
    "psutil>=5.9",
    "opencv-contrib-python>=4.9",
    "icecream>=2.1",
    "umap-learn>=0.5",
    # ``tabulate`` is the backend used by ``pandas.DataFrame.to_markdown``,
    # which the diagnostics report writer inlines for table previews.
    "tabulate>=0.9",
    # `utils_behavior` is a lab-internal package used throughout the codebase.
    # Install from GitHub: https://github.com/NeLy-EPFL/utils_behavior
    "utils_behavior",
]

[project.optional-dependencies]
interactive = [
    "bokeh>=3.4",
    "holoviews>=1.19",
    "panel>=1.4",
    "colorcet>=3.1",
    "shiny>=1.0",
]
video = [
    "moviepy>=2.0",
    "pygame>=2.5",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "black>=24.0",
    "ruff>=0.3",
    "pre-commit>=3.6",
]
all = [
    "ballpushing_utils[interactive,video,dev]",
]

[project.urls]
Homepage = "https://github.com/NeLy-EPFL/ballpushing_utils"
Repository = "https://github.com/NeLy-EPFL/ballpushing_utils"
Issues = "https://github.com/NeLy-EPFL/ballpushing_utils/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["ballpushing_utils*"]

[tool.setuptools.package-data]
ballpushing_utils = ["py.typed", "*.md"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = ["docs", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I", "UP"]
ignore = ["E501", "E722", "B008"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers"
testpaths = ["tests"]
markers = [
    "integration: integration tests (may require data on disk)",
    "slow: tests that take more than a few seconds",
    "unit: fast offline tests",
]

[tool.coverage.run]
source = ["src/ballpushing_utils"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
